provide a proper default for R2PipeSpawnOptions
The derived `Default` implementation for `R2PipeSpawnOptions` created an
empty `args`  and an empty `exepath`. That is almost
certainly not what was intended since the example that showed its use
looked like this:

```rust
let opts = R2PipeSpawnOptions {
        exepath: "radare2".to_owned(),
        ..Default::default()
    };
```

implying that the `Default` implementation would provide a reasonable
default and that you could just override the options that you wanted to
override. However, if you tried to do something like the following:

```rust
let opts = R2PipeSpawnOptions {
        args: vec!["-2"],
        ..Default::default()
    };
```

to just disable the warnings spat out on stderr, `R2Pipe` would fail to
spawn since the `exepath` was empty. This is the case since the `spawn`
function would default initialize the `exepath` if you didn't provide
_any_ `R2PipeSpawnOptions` by passing `None`, but simply took what you
passed in if you passed `Some(R2PipeSpawnOptions)`

By providing a proper `Default` implementation, this allows the previous
method of overriding the `args` for the options without having to also
override the `exepath`. It also simplifies the handling when you do not
provide any options.