handle errors with the initial null read
When the computer is busy, it can be the case that we get an error back
from `read_exact` if we are attempted to read the null byte before the
subprocess has fired up and connected to the stdout stream to write
anything. Instead of just returning an error in that case, we should
sleep a little while to see if the subprocess will start and emit our
initial null byte. By default, I am choosing to wait for up to 10 seconds
in 10ms intervals to see if we can connect without returning a failure.
However, there are two case where we can get a `UnexpectedEof`: the case
describe above is where the process is still coming up and hasn't
connected yet. The other case is when the process has started up,
but then exits unexpectedly for some reason. So, we want to check to see
if the process had exited and, if so, return that error to the user
instead.