remove unnecessary `mut` and `take` of `opts`
There was no reason to `take` out of the `Option<R2PipeSpawnOptions>` if
we are going to do an `unwrap_or_default`. I am not sure what possessed
me to write the code that way, but since I am touching adding an
additional option in a future commit, I wanted to go ahead and clean
that up.