catch an exited or aborted subprocess
If we get an `UnexpectedEof` from the `read_until` call, it is
worthwhile to report that as a separate error instead of requiring the
end-user to parse the IO error that comes back since, at this point, we
can go ahead and grab the exit status of the program and return that as
part of the error response. This will allow the end-user to make
decisions about whether to retry the request or abort if radare2
crashed.